fix missing samples
